+++ Common/src/main/java/org/example/dto/request/OperationFilterRequest.java
@@ -7,15 +7,15 @@ import java.time.Instant;
 
 @Getter
 public class OperationFilterRequest extends AuthorizedRequest {
-    private final Integer maxAmount;
-    private final Integer minAmount;
+    private final Double maxAmount;
+    private final Double minAmount;
     private final Instant dateFrom;
     private final Instant dateTo;
 
     public OperationFilterRequest(
             String token,
-            Integer maxAmount,
-            Integer minAmount,
+            Double maxAmount,
+            Double minAmount,
             Instant dateFrom,
             Instant dateTo
     ) {
