+++ Common/src/main/java/org/example/dto/request/OperationRequest.java
@@ -7,10 +7,10 @@ import java.time.Instant;
 
 @Getter
 public class OperationRequest extends AuthorizedRequest {
-    private final int amount;
+    private final double amount;
     private final Instant dateTime;
 
-    public OperationRequest(RequestAction action, String token, int amount, Instant dateTime) {
+    public OperationRequest(RequestAction action, String token, double amount, Instant dateTime) {
         super(action, token);
         this.amount = amount;
         this.dateTime = dateTime;
